+++ app/login/page.tsx
@@ -1,77 +1,70 @@
-// import React from 'react'
-
-// const Login = () => {
-//   return (
-//     <div className='bg-[#F1F1F1] p-0 w-full h-screen flex justify-end items-center'>
-//         <img src='' alt='' />
-
-//         <div className='bg-white flex w-1/2 h-2/3 p-8 m-24'>
-
-//         </div>
-//     </div>
-//   )
-// }
-
-// export default Login
-
-/**
- * This code was generated by Builder.io.
- */
 import React from "react";
-import LoginForm from "../../components/Form/LoginForm";
+import LoginForm from "../../components/Form/page";
 import Link from "next/link";
 import loginBanner from "../../assets/loginBanner.png"
 import Logo from "@/components/Logo/page";
+import userIcon from "../../assets/user.png";
+import passwordIcon from "../../assets/password.png";
+import { FormField } from "@/types/login";
+import { loginContent } from "@/constants/login";
 
-const loginContent = {
-    heading: "Welcome!",
-    subheading: "Please Log-in to get started!",
-    formFooterText: " Not a customer yet?",
-    formFooterLink: "Get started!"
-}
 
 const Login: React.FC = () => {
-  return (
-    <main className="overflow-hidden px-16 pt-14 pb-24 bg-[#F1F1F1] max-md:px-5 h-screen justify-center items-center flex font-libre">
-      <div className="flex gap-16 max-md:flex-col w-[90%]">
-        <section className="flex flex-col w-[56%] max-md:ml-0 max-md:w-full">
-          <div className="flex flex-col max-md:mt-10 max-md:max-w-full">
-            <Logo />
-            <img
-              loading="lazy"
-              src={loginBanner.src}
-              alt=""
-              className="object-fill self-end mt-4 max-w-full aspect-[1.41] max-md:mt-10 w-full h-full"
-            />
-          </div>
-        </section>
-        <section className="flex flex-col w-[44%] ml-5 max-md:ml-0 max-md:w-full">
-          <div className="flex  gap-8 overflow-hidden flex-col grow items-center justify-center px-14 pt-16 pb-9 mt-7 w-full bg-white max-md:px-5 max-md:mt-10 max-md:max-w-full">
-            <div className="flex flex-col gap-4">
-            <h1 className="ml-3 text-4xl font-semibold text-center text-black flex justify-center items-center">
-              {loginContent.heading}
-            </h1>
-            <p className="text-lg font-medium leading-5 text-center text-black">
-              {loginContent.subheading}
-            </p>
+    const formFields: FormField[] = [
+        { name: "username", placeholder: "Username or Email", type: "text", icon: userIcon.src },
+        { name: "password", placeholder: "Password", type: "password", icon: passwordIcon.src },
+    ];
+    return (
+        <main className="overflow-hidden px-16 pt-24 pb-24 bg-[#F1F1F1] max-md:px-5 h-screen justify-center items-center flex flex-col font-libre">
+            <div className="flex w-full items-start px-4">
+                <Logo />
             </div>
-            
-            <div>
-            <LoginForm />
-            {/* <SignupForm />
-            <ImageUpload /> */}
-           
-            <p className="mt-7 text-base font-medium leading-5 text-[#837a7a] flex items-center justify-center">
-              {loginContent.formFooterText} <span className="text-yellow-600">  <Link href='/register'>&nbsp;
-              {loginContent.formFooterLink}
-        </Link> </span><br />
-            </p>
+
+            <div className="flex lg:gap-16 md:gap-0 sm:gap-0 max-md:flex-col w-[90%] items-center">
+                <section className="flex-col w-[56%] justify-center items-center  hidden md:hidden lg:flex ">
+                    <div className="w-full">
+                        <div className="flex items-start"></div>
+                        <div className="flex justify-center items-center w-full h-full flex-grow">
+                            <img
+                                loading="lazy"
+                                src={loginBanner.src}
+                                alt=""
+                                className="object-contain max-w-full aspect-[1.41] w-full h-full hidden md:hidden lg:flex mt-4" // Hide image on medium screens
+                            />
+                        </div>
+                    </div>
+                </section>
+
+                <section className="flex flex-col lg:w-[44%] ml-5 md:w-full max-md:ml-0 max-md:w-full h-[80vh]">
+                    <div className="flex gap-8 overflow-hidden flex-col grow items-center justify-center px-14 pb-9 mt-7 w-full bg-white max-md:px-5 max-md:mt-10 max-md:max-w-full">
+                        <div className="flex flex-col gap-4">
+                            <h1 className="ml-3 text-4xl font-semibold text-center text-black flex justify-center items-center">
+                                {loginContent.heading}
+                            </h1>
+                            <p className="text-lg font-medium leading-5 text-center text-black">
+                                {loginContent.subheading}
+                            </p>
+                        </div>
+
+                        <div>
+                            <LoginForm formFields={formFields} />
+
+                            <p className="mt-7 text-base font-medium leading-5 text-[#837a7a] flex items-center justify-center">
+                                {loginContent.formFooterText}
+                                <span className="text-yellow-600">
+                                    <Link href="/register">
+                                        &nbsp;{loginContent.formFooterLink}
+                                    </Link>
+                                </span>
+                                <br />
+                            </p>
+                        </div>
+                    </div>
+                </section>
             </div>
-          </div>
-        </section>
-      </div>
-    </main>
-  );
+        </main>
+
+    );
 };
 
 export default Login;
