--- app/middlewareside/page.tsx
@@ -1,9 +1,0 @@
-"use client";
-
-const MiddlewareSide = () => {
-  return (
-    <div>Middleware</div>
-  )
-}
-
-export default MiddlewareSide
