--- components/InhabitrSignup/ImageUpload.tsx
@@ -1,34 +1,0 @@
-/**
- * This code was generated by Builder.io.
- */
-import React from "react";
-
-const ImageUpload: React.FC = () => {
-  return (
-    <div className="flex relative flex-col self-start mt-3 ml-12 max-w-full h-[33px] w-[179px] max-md:ml-2.5">
-      <div className="flex z-0 flex-col w-full rounded-md">
-        <div className="flex shrink-0 h-9 rounded-md" />
-      </div>
-      <label
-        htmlFor="imageUpload"
-        className="flex absolute bottom-1 left-3.5 z-0 gap-1 max-w-full h-6 text-base font-medium text-amber-300 uppercase w-[150px] cursor-pointer"
-      >
-        <img
-          loading="lazy"
-          src="https://cdn.builder.io/api/v1/image/assets/TEMP/f749dcfc45636fde5892c20520a7445af00fdfadadd6d57cb772581ee92fbefd?apiKey=446ff06a6e8649bda897294d130e767e&&apiKey=446ff06a6e8649bda897294d130e767e"
-          alt=""
-          className="object-contain shrink-0 w-6 aspect-square"
-        />
-        <span className="flex-auto my-auto">Upload Image</span>
-      </label>
-      <input
-        type="file"
-        id="imageUpload"
-        className="sr-only"
-        accept="image/*"
-      />
-    </div>
-  );
-};
-
-export default ImageUpload;
