--- components/InhabitrSignup/InhabitrSignup.tsx
@@ -1,52 +1,0 @@
-/**
- * This code was generated by Builder.io.
- */
-import React from "react";
-import Header from "../Logo/page";
-import SignupForm from "./SignupForm";
-import ImageUpload from "./ImageUpload";
-
-const InhabitrSignup: React.FC = () => {
-  return (
-    <main className="overflow-hidden px-16 pt-14 pb-24 bg-zinc-100 max-md:px-5">
-      <div className="flex gap-5 max-md:flex-col">
-        <section className="flex flex-col w-[56%] max-md:ml-0 max-md:w-full">
-          <div className="flex flex-col max-md:mt-10 max-md:max-w-full">
-            <Header />
-            <img
-              loading="lazy"
-              src="https://cdn.builder.io/api/v1/image/assets/TEMP/9de08e753d287a911fac8da25ef36f5e814bb47b8e8380376b37db8226ead3c3?apiKey=446ff06a6e8649bda897294d130e767e&&apiKey=446ff06a6e8649bda897294d130e767e"
-              alt=""
-              className="object-contain self-end mt-48 max-w-full aspect-[1.41] w-[625px] max-md:mt-10"
-            />
-          </div>
-        </section>
-        <section className="flex flex-col ml-5 w-[44%] max-md:ml-0 max-md:w-full">
-          <div className="flex overflow-hidden flex-col grow items-center px-14 pt-16 pb-9 mt-7 w-full bg-white max-md:px-5 max-md:mt-10 max-md:max-w-full">
-            <h1 className="ml-3 text-4xl font-semibold text-center text-black">
-              Let's get started!
-            </h1>
-            <p className="mt-3 text-lg font-medium leading-5 text-center text-black">
-              Signing up with Inhabitr's Ordering Portal <br /> is easier than
-              ever.
-            </p>
-            <p className="self-stretch mt-11 text-base font-medium leading-5 text-center text-black max-md:mt-10">
-              Complete the fields below and upload photos of your furniture. Our
-              AI will identify your furniture package for you
-            </p>
-            <SignupForm />
-            <ImageUpload />
-            <button className="px-16 py-3.5 mt-6 max-w-full text-base font-medium text-center text-black whitespace-nowrap bg-amber-300 rounded-3xl w-[216px] max-md:px-5">
-              SUBMIT
-            </button>
-            <p className="mt-7 text-base font-medium leading-5 text-yellow-600">
-              Learn more about Inhabitr AI <br />
-            </p>
-          </div>
-        </section>
-      </div>
-    </main>
-  );
-};
-
-export default InhabitrSignup;
