--- components/InhabitrSignup/SignupForm.tsx
@@ -1,42 +1,0 @@
-/**
- * This code was generated by Builder.io.
- */
-import React from "react";
-
-interface FormField {
-  name: string;
-  placeholder: string;
-  type: string;
-}
-
-const formFields: FormField[] = [
-  { name: "propertyName", placeholder: "Property Name", type: "text" },
-  { name: "website", placeholder: "Website", type: "url" },
-  { name: "contactName", placeholder: "Contact Name", type: "text" },
-  { name: "contactEmail", placeholder: "Contact Email", type: "email" },
-  { name: "contactNumber", placeholder: "Contact Number", type: "tel" },
-];
-
-const SignupForm: React.FC = () => {
-  return (
-    <form>
-      {formFields.map((field) => (
-        <div key={field.name} className="mt-3 w-full max-w-[317px]">
-          <label htmlFor={field.name} className="sr-only">
-            {field.placeholder}
-          </label>
-          <input
-            type={field.type}
-            id={field.name}
-            name={field.name}
-            placeholder={field.placeholder}
-            className="px-5 py-5 w-full text-xs font-medium bg-white rounded-md text-stone-500"
-            required
-          />
-        </div>
-      ))}
-    </form>
-  );
-};
-
-export default SignupForm;
