+++ constants/login.ts
@@ -1,0 +1,7 @@
+export const loginContent = {
+    heading: "Welcome!",
+    subheading: "Please Log-in to get started!",
+    formFooterText: " Not a customer yet?",
+    formFooterLink: "Get started!"
+}
+
