+++ package.json
@@ -10,6 +10,7 @@
   "dependencies": {
     "@reduxjs/toolkit": "^2.2.7",
     "next": "14.2.5",
+    "next-router-mock": "^0.9.13",
     "react": "^18",
     "react-dom": "^18",
     "react-redux": "^9.1.2"
