+++ types/login.ts
@@ -1,0 +1,10 @@
+export interface FormField {
+    name: string;
+    placeholder: string;
+    type: string;
+    icon: string;
+}
+
+export interface LoginFormProps {
+    formFields: FormField[];
+}
